benchbboxtest/datasets/image/celebamask.py
import os
import numpy as np
from typing import Dict, List, Tuple, Optional, Union
import torch
from torch.utils.data import Dataset, DataLoader
from PIL import Image
import requests
import zipfile
import io
from ...core import DataGenerator


class CelebAMaskDataset(Dataset):
    """
    Dataset class for CelebAMask-HQ dataset.
    """
    
    def __init__(self, root_dir: str, transform=None):
        """
        Initialize the CelebAMask-HQ dataset.
        
        Args:
            root_dir: Root directory containing the dataset
            transform: Optional transform to be applied to the images
        """
        self.root_dir = root_dir
        self.transform = transform
        self.image_dir = os.path.join(root_dir, 'CelebA-HQ-img')
        self.mask_dir = os.path.join(root_dir, 'CelebAMask-HQ-mask-anno')
        
        # Get list of image files
        self.image_files = sorted([f for f in os.listdir(self.image_dir) 
                                  if f.endswith('.jpg') or f.endswith('.png')])
    
    def __len__(self):
        return len(self.image_files)
    
    def __getitem__(self, idx):
        img_name = os.path.join(self.image_dir, self.image_files[idx])
        image = Image.open(img_name).convert('RGB')
        
        # Get corresponding mask files
        img_id = self.image_files[idx].split('.')[0]
        mask_files = [f for f in os.listdir(self.mask_dir) if f.startswith(img_id)]
        
        # Load masks
        masks = {}
        for mask_file in mask_files:
            region_name = mask_file.split('_')[1]
            mask_path = os.path.join(self.mask_dir, mask_file)
            mask = Image.open(mask_path).convert('L')
            masks[region_name] = np.array(mask)
        
        if self.transform:
            image = self.transform(image)
        
        return {'image': image, 'masks': masks, 'id': img_id}


def download_celebamask_hq(target_dir: str):
    """
    Download and extract the CelebAMask-HQ dataset.
    
    Args:
        target_dir: Directory to save the dataset
    """
    # This is a placeholder - in a real implementation, you would need to:
    # 1. Download the dataset from the official source
    # 2. Extract it to the target directory
    # 3. Organize the files as needed
    
    os.makedirs(target_dir, exist_ok=True)
    print(f"CelebAMask-HQ dataset would be downloaded to {target_dir}")
    print("Note: This is a placeholder. In a real implementation, you would need to download the dataset from the official source.")
    print("The CelebAMask-HQ dataset is available at: https://github.com/switchablenorms/CelebAMask-HQ")


def mask_region(image: np.ndarray, segmentation_mask: np.ndarray, 
                value: Union[int, Tuple[int, int, int]] = 0) -> np.ndarray:
    """
    Apply a mask to a specific region of an image.
    
    Args:
        image: Input image (H, W, C)
        segmentation_mask: Binary mask for the region (H, W)
        value: Value to fill the masked region with
        
    Returns:
        Masked image
    """
    masked_image = image.copy()
    
    if len(image.shape) == 3:  # RGB image
        for c in range(image.shape[2]):
            masked_image[:, :, c] = np.where(segmentation_mask > 0, value if isinstance(value, int) else value[c], image[:, :, c])
    else:  # Grayscale image
        masked_image = np.where(segmentation_mask > 0, value, image)
    
    return masked_image


class CelebAMaskGenerator(DataGenerator):
    """
    Data generator for CelebAMask-HQ dataset.
    
    Implements data generation for conditional independence testing using facial attributes and regions.
    """
    
    # Mapping from attributes to facial regions
    ATTRIBUTE_REGION_MAP = {
        'Narrow_Eyes': 'eye',
        'Pointy_Nose': 'nose',
        'Big_Lips': 'mouth',
        'Bushy_Eyebrows': 'eyebrow',
        'Male': 'face',
        'Smiling': 'mouth',
        'Wearing_Earrings': 'ear',
        'Wearing_Hat': 'hair',
        'Wearing_Lipstick': 'mouth',
        'Wearing_Necklace': 'neck'
    }
    
    def __init__(self, dataset_path: str, attribute: str, crucial_region: str = None):
        """
        Initialize the CelebAMask-HQ generator.
        
        Args:
            dataset_path: Path to the CelebAMask-HQ dataset
            attribute: Facial attribute to predict (Y)
            crucial_region: Crucial region for the attribute (if None, determined from ATTRIBUTE_REGION_MAP)
        """
        self.dataset_path = dataset_path
        self.attribute = attribute
        
        if crucial_region is None and attribute in self.ATTRIBUTE_REGION_MAP:
            self.crucial_region = self.ATTRIBUTE_REGION_MAP[attribute]
        else:
            self.crucial_region = crucial_region
    
    def generate_null(self, n_samples: int, **kwargs) -> Dict[str, np.ndarray]:
        """
        Generate data under the null hypothesis (X ⊥ Y | Z).
        
        For the null hypothesis, we mask non-crucial regions, making X and Y conditionally independent given Z.
        
        Args:
            n_samples: Number of samples to generate
            **kwargs: Additional parameters
            
        Returns:
            Dictionary containing 'X', 'Y', and 'Z' arrays
        """
        # This is a simplified implementation
        # In a real implementation, you would:
        # 1. Load n_samples images from the dataset
        # 2. Extract the attribute values (Y)
        # 3. Extract the crucial region features (Z)
        # 4. Mask non-crucial regions to create X
        
        # Placeholder implementation
        X = np.random.rand(n_samples, 64, 64, 3)  # Masked images (non-crucial regions)
        Y = np.random.randint(0, 2, size=n_samples)  # Binary attribute values
        Z = np.random.rand(n_samples, 32)  # Features from crucial regions
        
        return {'X': X, 'Y': Y.reshape(-1, 1), 'Z': Z}
    
    def generate_alternative(self, n_samples: int, **kwargs) -> Dict[str, np.ndarray]:
        """
        Generate data under the alternative hypothesis (X ⊥̸ Y | Z).
        
        For the alternative hypothesis, we mask crucial regions, making X and Y conditionally dependent given Z.
        
        Args:
            n_samples: Number of samples to generate
            **kwargs: Additional parameters
            
        Returns:
            Dictionary containing 'X', 'Y', and 'Z' arrays
        """
        # This is a simplified implementation
        # In a real implementation, you would:
        # 1. Load n_samples images from the dataset
        # 2. Extract the attribute values (Y)
        # 3. Extract features from non-crucial regions (Z)
        # 4. Mask crucial regions to create X
        
        # Placeholder implementation
        X = np.random.rand(n_samples, 64, 64, 3)  # Masked images (crucial regions)
        Y = np.random.randint(0, 2, size=n_samples)  # Binary attribute values
        Z = np.random.rand(n_samples, 32)  # Features from non-crucial regions
        
        return {'X': X, 'Y': Y.reshape(-1, 1), 'Z': Z} 