requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.10.0
torchvision>=0.11.0
pillow>=8.0.0
transformers>=4.15.0
arxiv>=1.4.2
requests>=2.25.0
tqdm>=4.62.0
pytest>=6.2.5
